requirements.txt
websockets==11.*
Jinja2==3.*
uvloop==0.15.3
japronto==0.1.2
boto3==1.26.*
psycopg2-binary==2.9.*
requests==2.*
css-html-js-minify==2.5.5
uvicorn==0.21.*
gunicorn==20.*
pydantic==1.*
python-dotenv==0.*
tortoise-orm==0.19.*
asyncpg==0.27.*
aerich==0.7.*
# sqlalchemy==1.3.*
# sqlalchemy-migrate==0.13.*
# tensorflow-cpu==1.15.0
# stable_baselines==2.10.1
# numpy
